requirements.txt
asttokens==2.4.1
attrs==23.2.0
bleach==6.1.0
bokeh==3.3.2
Cartopy==0.22.0
cdsapi==0.6.1
certifi==2023.11.17
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colorcet==3.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fiona==1.9.5
fonttools==4.47.0
geopandas==0.14.2
geoviews==1.11.0
holoviews==1.18.1
hvplot==0.9.1
idna==3.6
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyter-bokeh==3.0.7
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
linkify-it-py==2.0.2
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
nest-asyncio==1.5.8
netCDF4==1.6.5
numpy==1.26.3
OWSLib==0.29.3
packaging==23.2
pandas==2.1.4
panel==1.3.6
param==2.0.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pyct==0.5.0
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyviz_comms==3.0.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.9
xarray==2023.12.0
xyzservices==2023.10.1
